+++ src/WinOldRecovery.Core/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace WinOldRecovery.Core;
-
-public class Class1
-{
-
-}