+++ src/WinOldRecovery.Native/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace WinOldRecovery.Native;
-
-public class Class1
-{
-
-}